requirements.txt
pymongo==4.7.2
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
requests==2.32.3
